etl/requirements.txt
elasticsearch==7.17.6
python-dotenv==0.21.0
psycopg2-binary==2.9
pydantic==1.10.2